Move Profile Screen location to proper package
